Add layout and goal details components; refactor imports and navigation
