+++ app/goals/[id].tsx
@@ -1,0 +1,8 @@
+import { Text, View } from 'react-native';
+export default function GoalDetails() {
+    return (
+        <View>
+            <Text>This is details of Goal..</Text>
+        </View>
+    );
+}
